# specs/002-schedule-mode/contracts/api.yaml
openapi: 3.0.0
info:
  title: ViMax Web 调度模式与调度管理 API
  version: 1.0.0
  description: >
    调度模式与调度管理功能的后端 API 契约。
    (A) Web 端 API：backend 暴露给 frontend 的新增端点。
    (B) 调度服务对接契约：backend 调用外部 Vimax Scheduler 的端点映射（X-API-Key 鉴权）。
servers:
  - url: http://127.0.0.1:8000

# =====================================================================
# (A) Web 端 API（backend → frontend）
# =====================================================================
paths:
  /api/projects/{project_id}/schedule-mode:
    patch:
      operationId: setScheduleMode
      summary: 开启/关闭项目调度模式
      description: >
        开启后项目进入调度模式。关闭时若存在未结束调度任务须先取消（由调用方确认）。
      parameters:
        - name: project_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [schedule_mode]
              properties:
                schedule_mode: { type: boolean, description: true=开启调度模式 }
      responses:
        '200':
          description: 项目调度字段已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSchedule'
        '400':
          description: 关闭调度模式但存在未结束调度任务（提示需先取消）
        '503':
          description: 调度服务不可用
  /api/projects/{project_id}/schedule-status:
    get:
      operationId: getScheduleStatus
      summary: 获取项目调度状态
      description: 返回本地调度状态与调度服务侧实时统计（供"调度中"判定与模式切换约束）。
      parameters:
        - name: project_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 调度状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSchedule'
...
components:
  schemas:
    ProjectSchedule:
      type: object
      properties:
        project_id: { type: integer }
        schedule_mode: { type: boolean }
        schedule_status: { type: string, enum: [idle, scheduling] }
        pending_tasks: { type: integer, description: 未结束调度任务数（本地队列 + 调度服务 stats.pending） }
        can_switch_to_execute: { type: boolean, description: 是否允许切回执行模式 }
...